requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
outcome==1.3.0.post0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
PySocks==1.7.1
requests~=2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
SpeechRecognition==3.10.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
selenium~=4.21.0
python-dotenv~=1.0.1
fitz~=0.0.1.dev2